requirements.txt
Pillow>=10.2.0
numpy>=1.26.4
ExifRead>=3.0.0
PyPDF2>=3.0.1
python-docx>=1.1.0
openpyxl>=3.1.2